Rename and reposition file to avoid same name
